src/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1